requirements.txt
tzlocal==1.4
requests==2.9.1
requests-oauthlib==0.8.0
Django==1.11.5
django-allauth==0.34.0
django-address==0.1.5
django-model-utils==3.0.0
django-notifications-hq==1.3
django-static-precompiler==1.6
django-storages==1.6.5
python-openid==2.2.5
python-utils==2.2.0
numpy==1.11.0
numpy-stl==2.3.2
Pillow==4.3.0
olefile==0.44
boto3==1.4.7
botocore==1.7.1
